requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
openai
python-dotenv
sqlalchemy
pgvector
pdfplumber